web/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=12.0
python-dotenv>=1.0.1
pydantic>=2.0.0
